refactor: move single reservation page
