+++ apps/admin-ui/src/component/my-units/ReservationPopupContent.tsx
@@ -1,15 +1,14 @@
 import React from "react";
 import { Link } from "react-router-dom";
 import styled from "styled-components";
-import { ReservationType } from "common/types/gql-types";
+import type { ReservationType } from "common/types/gql-types";
 import { Permission } from "@modules/permissionHelper";
 import { reservationUrl } from "@common/urls";
-import { formatTime } from "@common/util";
+import { getReserveeName } from "@common/util";
+import { formatTime } from "@/helpers";
 import { DenseVerticalFlex } from "@styles/layout";
 import { CELL_BORDER } from "./const";
 import VisibleIfPermission from "@component/VisibleIfPermission";
-// TODO this is imported from incorrect place (move them to common)
-import { getReserveeName } from "@/spa/reservations/requested/util";
 
 const PopupContent = styled.div`
   border: ${CELL_BORDER};
