+++ apps/admin-ui/src/spa/reservations/[id]/index.tsx
